@@ -36,10 +36,12 @@ import {
 import {
     expandTo18Decimals,
     MaxUint256,
-} from './shared/utilities.js'
+    events,
+    emits,
+} from './shared/utilities'
 
-const TOTAL_SUPPLY = BigInt(expandTo18Decimals( 10000 ))
-const TEST_AMOUNT = BigInt(expandTo18Decimals( 10 ))
+const TOTAL_SUPPLY = BigInt( expandTo18Decimals( 10000 ) )
+const TEST_AMOUNT = BigInt( expandTo18Decimals( 10 ) )
 
 const wallet = {
     ...WALLETS[0],
@@ -74,12 +76,24 @@ describe( 'AEX9', () => {
         expect( await exe( x => x.balance_str( wallet.address ) ) ).to.eq( TOTAL_SUPPLY.toString() )
     } )
     it( 'approve', async () => {
-        await contract.exe( x => x.create_allowance( other.address, TEST_AMOUNT) )
+        await contract.exe(
+            x => x.create_allowance( other.address, TEST_AMOUNT ),
+            events( 
+                emits( "Allowance" ).withArgs(
+                    '3ctqe1KNTz5XFTByw',
+                    'tWZrf8ehmY7CyB1JAoBmWJEeThwWnDpU4NadUdzxVSbzDgKjP',
+                    '39519965516565108473327470053407124751867067078530473195651550649472681599133',
+                )
+            )
 
-        expect( await contract.exe( x => x.allowance_unfolded(
-            wallet.address,
-            other.address
-        ) ) ).to.eq( TEST_AMOUNT )
+        )
+
+        expect( await contract.exe(
+            x => x.allowance_unfolded(
+                wallet.address,
+                other.address
+            ),
+        ) ).to.eq( TEST_AMOUNT )
     } )
     it( 'transfer', async () => {
         await contract.exe( x => x.transfer( other.address, TEST_AMOUNT ) )
@@ -125,7 +139,7 @@ describe( 'AEX9', () => {
         expect( ( await contract.exe( x => x.allowance_unfolded(
             wallet.address,
             other.address
-        ) ) ).toString() ).to.eq( (BigInt(MaxUint256) - TEST_AMOUNT).toString() )
+        ) ) ).toString() ).to.eq( ( BigInt( MaxUint256 ) - TEST_AMOUNT ).toString() )
         expect( await contract.exe( x => x.balance( wallet.address ) ) ).to.eq(
             TOTAL_SUPPLY - TEST_AMOUNT
         )
@@ -134,7 +148,7 @@ describe( 'AEX9', () => {
         )
     } )
     it( 'permit', async () => {
-        const deadline = BigInt(MaxUint256)
+        const deadline = BigInt( MaxUint256 )
 
         await contract.exe( x => x.permit(
             wallet.address,
